CS-493 (Cloud Application Development)/requirements.txt
Flask==3.0.0
google-cloud-datastore==2.15.1
python-jose
python-dotenv
authlib
protobuf